Moves currentDukeDsUser functionality to a new application controller
- Other controllers can inject this controller to look up current user
