+++ app/templates/deliveries/index.hbs
@@ -1,4 +1,5 @@
 {{delivery-breadcrumbs selectedRouteName="deliveries.index"}}
 <h1>Deliveries</h1>
+
 {{delivery-table model}}
