+++ tests/unit/controllers/deliveries/show/index-test.js
@@ -2,7 +2,7 @@ import { moduleFor, test } from 'ember-qunit';
 
 moduleFor('controller:deliveries/show/index', 'Unit | Controller | deliveries/show/index', {
   // Specify the other units that are required for this test.
-  needs: ['service:session', 'service:duke-ds-user', 'model:duke-ds-user']
+  needs: ['controller:application']
 });
 
 test('it is ok', function(assert) {
